moved model settings files into single model_settings folder organized settings for caffenet added files to ignore list removed ugb since it doesn't work

+++ build_default.sh
@@ -1,5 +1,9 @@
+#!/usr/bin/env bash
 cd caffe
-cp Makefile.config.example Makefile.config
+# if Makefile.config has already existed, then don't overwrite it
+if [ ! -e  Makefile.config ]; then
+    cp Makefile.config.example Makefile.config
+fi
 make all pycaffe
 cd ..
 cp settings_user.py.example settings_user.py
